auth: Support multiple content types in webserver responses
This patch modifies the response body mechanism in the webserver
to no longer assume JSON output, but allow handlers to specify
JSON, YAML, or plain text. It also adds detection of YAML
support in the request so that handlers can choose which type
to send in their response. Since there is not yet a standard
MIME type for YAML, 'application/x-yaml' is used since it
appears to be the most commonly used type.

Finally, one error path chose to send plain text output instead
of JSON even when the request indicated JSON support; this has
been reversed so that JSON has priority as is done in other
locations in the code.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>