Add a potential method to allow Sphinx to consume API spec file
If the API spec file uses YAML 'merge keys', the current version
of the OpenAPI parser doesn't understand them and the build of
the documentation fails.

One option to resolve this is to use the Python YAML module to
'expand' the merge keys into regular anchors and aliases which
are supported by the OpenAPI parser. This patch uses that method
to produce a temporary version of the spec file for Sphinx to
read.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>