Allow contributors to talk to check-spelling in their own forks
At the present this feature only works if the base and head branch of the PR are in the same repository.

See: https://github.com/check-spelling/check-spelling/wiki/Feature:-Update-expect-list#read-write-deploy-key-limitations

Given this restriction, it seems that enabling it for contributors in forks would enable users to use the feature.

The limitation should reduce concerns about security to the named organization repository instance.